refactor: split coarse contract and server boundary modules #261
